// libs/shared/ui/button/src/lib/models/button.models.ts
export type ButtonColor =
  | 'primary'
  | 'secondary-1'
  | 'secondary-2'
  | 'success'
  | 'disruption';

export type ButtonType = 'filled' | 'link' | 'outline';

export type ButtonSize = 'x-small' | 'small' | 'medium' | 'large' | 'x-large';

export type ButtonShape = 'square' | 'rounded';

export type ButtonDisplay = 'inline-block' | 'inline' | 'block' | 'icon-only';

export type ButtonClass = 'floating';

export type ButtonBehaviour = 'button' | 'submit' | 'reset';